requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-ollama==0.2.1
langchain-huggingface==0.1.2
faiss-cpu==1.9.0.post1
rank-bm25==0.2.2
sentence-transformers==3.3.1
transformers==4.46.3
einops==0.8.0
fastapi==0.115.5
uvicorn==0.32.1
streamlit==1.40.2
pypdf==5.1.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.10.1